Se elimino el controlador de Autenticacion duplicado, metodos duplicados en user service y se añadio la intefaz del helper del token. Por ultimo se registraron los servicios de las imagenes en el startup
