+++ OngProject/Core/Interfaces/IJwtHelper.cs
@@ -1,0 +1,9 @@
+﻿using OngProject.Entities;
+
+namespace OngProject.Core.Interfaces
+{
+    public interface IJwtHelper
+    {
+        string GenerateJwtToken(User user);
+    }
+}
