+++ OngProject/Startup.cs
@@ -59,9 +59,11 @@ namespace OngProject
             services.AddScoped<ITestimonialsService, TestimonialsService>();
             services.AddScoped<IUserService, UsersService>();
             services.AddScoped<ISlideSerivice, SlideService>();
+            services.AddScoped<IImageService, ImageService>();
+            services.AddScoped<IS3AwsHelper, S3AwsHelper>();
             services.AddScoped<IContactService, ContactService>();
             services.Configure<AuthMessageSenderOptions>(Configuration);
-            services.AddScoped<JwtHelper>();
+            services.AddScoped<IJwtHelper, JwtHelper>();
 
 
             // JWT Token Generator
